it is float, not int

+++ src/Ads/Widgets/RelatedArticles.php
@@ -4,7 +4,7 @@ namespace BackendApp\Ads\Widgets;
 
 class RelatedArticles implements WidgetInterface
 {
-	public function getPointsValue(array $widget) : int
+	public function getPointsValue(array $widget) : float
 	{
 		return 1;
 	}
